apps/api/requirements.txt
# ─── UMS API Dependencies ──────────────────────────────────
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.3

# Security
PyJWT==2.9.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
greenlet==3.1.1

# Redis (for chat pub/sub & caching)
redis==5.1.1

# Utilities
python-multipart==0.0.12
httpx==0.27.2

# Dev & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
ruff==0.6.9
